requirements.txt
Flask==3.1.1
Flask-CORS==6.0.0
Flask-SQLAlchemy==3.1.1
Pillow==10.4.0
pytesseract==0.3.13
openai==1.58.1
python-dotenv==1.0.1
requests==2.32.3
pdf2image==1.17.0
python-magic==0.4.27
Werkzeug==3.1.3

